refactor: reorganize project structure by moving KMA response
